refactor: CheatSheetのロジックをカスタムフックへ抽出（フェーズ2）
データロード/DnD/別ウィンドウ編集/CRUD/編集モードライフサイクル/
検索スクロールを hooks/cheat-sheet 配下の6フックへ切り出し。
CheatSheet.tsx は各フックを組み立てるオーケストレーターとなり
602行→362行（元1490行）に縮小。

- useCheatSheetData: タイトル/データのロード・reload購読
- useEditBlockDnd: 編集ブロックの並べ替えDnD
- useEditBlockActions: コマンド/グループのCRUD
- useEditWindows: edit_command/edit_group ウィンドウ開閉と保存受信
- useCheatSheetEditMode: 編集モードの開始/破棄/保存・確認ダイアログ
- useScrollToCommand: 検索からのスクロール/フォーカス
